brayanAP

+++ FingerNationAppMacro/FingerNationAppMacro/models/ModFingerNation.cs
@@ -38,8 +38,8 @@ namespace FingerNationAppMacro.models
         public int id { get; set; }
         public string nombre { get; set; }
         public int edad { get; set; }
-        public float altura { get; set; }
-        public float peso { get; set; }
+        public double altura { get; set; }
+        public double peso { get; set; }
         public string sexo { get; set; }
     }
